+++ jsonx/src/main/java/com/github/panpf/jsonx/ToBeanOrNull.java
@@ -23,5 +23,5 @@ import org.json.JSONObject;
 
 public interface ToBeanOrNull<T> {
     @Nullable
-    T toBean(@NotNull JSONObject jsonObject) throws JSONException;
+    T toBeanOrNull(@NotNull JSONObject jsonObject) throws JSONException;
 }