+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## new
+* change: ToBeanOrNull.toBean() rename to toBeanOrNull()
+
 ## v1.0.2
 * change: The parameters of ToBeanOrNull and ToJSONObjectOrNull are now marked as NotNull
 * new: Added toBeanArray() and toStringArrayList() methods